fix(protocol-designer): remove dropdown menu for module quantity (#17531)
* fix(protocol-designer): remove dropdown menu for module quantity
